+++ src/basta/constants.py
@@ -81,15 +81,18 @@ class SeismicfitAliases:
         "kbcd08": "KBC08",
         "kjeldsen08": "KBC08",
         "kjeldsen2008": "KBC08",
-        "two-term-bg14": "two-term-BG14",
-        "twotermbg14": "two-term-BG14",
-        "bg14": "two-term-BG14",
-        "ball14": "two-term-BG14",
-        "ball2014": "two-term-BG14",
-        "cubic-term-bg14": "cubic-term-BG14",
-        "cubicbg14": "cubic-term-BG14",
-        "cubic-bg14": "cubic-term-BG14",
-        "cbg14": "cubic-term-BG14",
+        "two-term-bg14": "two_term_BG14",
+        "two_term_bg14": "two_term_BG14",
+        "twotermbg14": "two_term_BG14",
+        "bg14": "two_term_BG14",
+        "ball14": "two_term_BG14",
+        "ball2014": "two_term_BG14",
+        "cubic-term-bg14": "cubic_term_BG14",
+        "cubic_term_bg14": "cubic_term_BG14",
+        "cubicbg14": "cubic_term_BG14",
+        "cubic-bg14": "cubic_term_bg14",
+        "cubic_bg14": "cubic_term_bg14",
+        "cbg14": "cubic_term_BG14",
     }
